requirements.txt
# Core HTTP & Scraping
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.5
aiofiles==23.2.1
beautifulsoup4==4.12.3
lxml==5.2.1
selenium==4.21.0
playwright==1.44.0

# Data Processing
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
fastparquet==2024.2.0
openpyxl==3.1.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1

# Async
asyncio-throttle==1.0.2
tenacity==8.3.0

# Data Quality
great_expectations==0.18.15
pydantic==2.7.4
pandera==0.19.3

# Scheduling & Orchestration
apscheduler==3.10.4
prefect==2.19.7

# Hashing & Security
cryptography==42.0.8
bcrypt==4.1.3
python-jose==3.3.0

# Logging & Monitoring
structlog==24.2.0
prometheus-client==0.20.0
sentry-sdk==2.5.1

# Utilities
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
tqdm==4.66.4
humanize==4.9.0
python-dateutil==2.9.0
pytz==2024.1
chardet==5.2.0
pyyaml==6.0.1

# XML Processing
xmltodict==0.13.0
defusedxml==0.7.1

# ZIP Handling
zipfile38==0.0.3

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0
responses==0.25.3
factory-boy==3.3.0

# Documentation
sphinx==7.3.7
